Fixed issue of ball getting "stuck" to paddle
Separated collision checker into two parts, top and bottom.

Uses assumption that if the bottom is colliding with something, we
ALWAYS want vy to become positive, and if the top is colliding, we
ALWAYS want vy to become negative.

This solves the problem of the ball being able to become stuck with
it's hitbox straddling the paddle.